+++ frontend/src/pages/workspaces/[workspaceSlug]/templates/[templateCode]/code.tsx
@@ -68,7 +68,7 @@ export const getServerSideProps = createGetServerSideProps({
   requireAuth: true,
   async getServerSideProps(ctx, client) {
     await FilesEditor.prefetch(ctx, client);
-    
+
     const originalProps = createTemplatePageServerSideProps();
     return await originalProps.getServerSideProps(ctx, client);
   },